back/requirements.txt
fastapi==0.115.0
SQLAlchemy==2.0.35
paho-mqtt==2.1.0
annotated-types==0.7.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
typing_extensions==4.12.2
cryptography==43.0.1
bcrypt==4.2.0
PyJWT